Info

+++ scripts/cleanup.sh
@@ -1,18 +1,23 @@
-#!/bin/bash -xv
+#!/bin/bash
 # clean all phpcompta related DB.
 DOMAIN="testing"
 export PGPASSWORD="dany"
 export PGUSER="phpcompta"
 export PGHOST=localhost
-dropdb  ${DOMAIN}account_repository
-dropdb  ${DOMAIN}dossier1
-dropdb   ${DOMAIN}dossier3
-dropdb   ${DOMAIN}dossier4
-dropdb   ${DOMAIN}dossier5
-dropdb   ${DOMAIN}dossier13
+echo "Etes vous sur de vouloir effacer les db du domaine $DOMAINE Y/N ?"
+read A
+if [ "$A" == 'Y' ];then
+    dropdb  ${DOMAIN}account_repository
+    dropdb  ${DOMAIN}dossier1
+    dropdb   ${DOMAIN}dossier3
+    dropdb   ${DOMAIN}dossier4
+    dropdb   ${DOMAIN}dossier5
+    dropdb   ${DOMAIN}dossier13
 
-dropdb   ${DOMAIN}mod1
-dropdb   ${DOMAIN}mod2
-dropdb   ${DOMAIN}mod3
-dropdb   ${DOMAIN}mod7
-
+    dropdb   ${DOMAIN}mod1
+    dropdb   ${DOMAIN}mod2
+    dropdb   ${DOMAIN}mod3
+    dropdb   ${DOMAIN}mod7
+else 
+    echo "Effacement annule"
+fi
